redirect to HOME on user logout
