+++ applications/fishing-map/src/features/sidebar/Sidebar.tsx
@@ -1,5 +1,9 @@
 import React, { Suspense, useMemo } from 'react'
-import { useSelector } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
+import Spinner from '@globalfishingwatch/ui-components/dist/spinner'
+import Button from '@globalfishingwatch/ui-components/dist/button'
+import { logoutUserThunk } from 'features/user/user.slice'
+import { isUserAuthorized, isUserLogged } from 'features/user/user.selectors'
 import Search from 'features/search/Search'
 import { selectSearchQuery } from 'features/app/app.selectors'
 import { selectLocationType } from 'routes/routes.selectors'
@@ -18,17 +22,38 @@ type SidebarProps = {
 function Sidebar({ onMenuClick }: SidebarProps) {
   const searchQuery = useSelector(selectSearchQuery)
   const locationType = useSelector(selectLocationType)
+  const userLogged = useSelector(isUserLogged)
+  const userAuthorized = useSelector(isUserAuthorized)
+  const dispatch = useDispatch()
 
   const sidebarComponent = useMemo(() => {
+    if (!userLogged) {
+      return <Spinner />
+    }
+    // TODO remove once public release and permissions to use map in anonymous user
+    if (!userAuthorized) {
+      return (
+        <div className={styles.placeholder}>
+          <h2>We're sorry but your user is not authorized to use this app yet</h2>
+          <Button
+            onClick={async () => {
+              dispatch(logoutUserThunk({ redirectToLogin: true }))
+            }}
+          >
+            Logout
+          </Button>
+        </div>
+      )
+    }
+
     if (locationType === USER) {
       return <User />
     }
     if (locationType === WORKSPACES_LIST) {
       return <WorkspacesList />
     }
-    // TODO: show loading when datasets and dataviews pending
     return <Workspace />
-  }, [locationType])
+  }, [dispatch, locationType, userAuthorized, userLogged])
 
   if (searchQuery !== undefined) {
     return <Search />
@@ -39,7 +64,7 @@ function Sidebar({ onMenuClick }: SidebarProps) {
         <CategoryTabs onMenuClick={onMenuClick} />
         <div className="scrollContainer">
           <SidebarHeader />
-          {sidebarComponent}
+          {userLogged ? sidebarComponent : <Spinner />}
         </div>
       </div>
     </Suspense>
