+++ applications/fishing-map/src/features/user/user.selectors.ts
@@ -1,5 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit'
 import { selectWorkspaces } from 'features/workspaces-list/workspaces-list.slice'
+import { AsyncReducerStatus } from 'types'
 import { selectUserStatus, selectUserLogged, selectUserData, GUEST_USER_TYPE } from './user.slice'
 
 export const isGuestUser = createSelector([selectUserData], (userData) => {
@@ -9,7 +10,7 @@ export const isGuestUser = createSelector([selectUserData], (userData) => {
 export const isUserLogged = createSelector(
   [selectUserStatus, selectUserLogged],
   (status, logged) => {
-    return status === 'finished' && logged
+    return status === AsyncReducerStatus.Finished && logged
   }
 )
